+++ packages/ko/src/core/config.ts
@@ -65,6 +65,7 @@ class Config {
         disableLazyImports: false,
       },
       autoPolyfills: false,
+      logLevel: 'error',
     };
   }
 }