+++ packages/ko/src/webpack/index.ts
@@ -63,6 +63,17 @@ class WebpackConfig {
     return require(pkgPath).version;
   }
 
+  get stats() {
+    const mapping: Record<NonNullable<IOptions['logLevel']>, string> = {
+      info: 'normal',
+      error: 'errors-only',
+      warn: 'errors-warnings',
+      none: 'none',
+    };
+    const { logLevel } = this.opts;
+    return mapping[logLevel!] as Configuration['stats'];
+  }
+
   get base() {
     const { cwd, publicPath, entry, outputPath, alias, hash, analyzer } =
       this.opts;
@@ -112,8 +123,9 @@ class WebpackConfig {
         hints: false,
       },
       cache: this.cache,
-      stats: {
-        cachedModules: false,
+      stats: this.stats,
+      infrastructureLogging: {
+        level: 'error',
       },
     };
     return webpackBaseConf as Configuration;